requirements.txt
asgiref==3.5.2
boto3==1.24.91
botocore==1.27.91
dj-database-url==1.0.0
Django==4.0.6
django-ckeditor==6.4.2
django-filter==22.1
django-js-asset==2.0.0
django-multiselectfield==0.1.12
django-storages==1.13.1
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.12
whitenoise==6.2.0
